[#15] feat(Comment): comment entity 생성 및 User 관련 entity 수정
