+++ src/main/java/kr/sparta/rchive/domain/comment/entity/Comment.java
@@ -1,4 +1,4 @@
-package kr.sparta.rchive.domain.user.entity;
+package kr.sparta.rchive.domain.comment.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,15 +7,14 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import kr.sparta.rchive.domain.user.enums.GenderEnum;
+import kr.sparta.rchive.domain.educationData.entity.EducationData;
+import kr.sparta.rchive.domain.user.entity.Track;
+import kr.sparta.rchive.domain.user.entity.User;
 import kr.sparta.rchive.domain.user.enums.OauthTypeEnum;
-import kr.sparta.rchive.domain.user.enums.TrackEnum;
 import kr.sparta.rchive.global.entity.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -25,22 +24,21 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
 
 @Entity
-@Table(name="tb_track")
+@Table(name="tb_comment")
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Track extends BaseTimeEntity {
+public class Comment extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 20)
-    @Enumerated(value = EnumType.STRING)
-    private TrackEnum track;
+    @Column
+    private String parentsCommentId;
 
     @Column(nullable = false)
-    private int period;
+    private String content;
 
     @Column(nullable = false)
     @ColumnDefault("false")
@@ -49,7 +47,11 @@ public class Track extends BaseTimeEntity {
     @Column
     private LocalDateTime deletedAt;
 
-    @OneToMany(mappedBy = "track")
-    List<Role> roleList = new ArrayList<>();
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User user;
 
+    @ManyToOne
+    @JoinColumn(name = "education_data_id")
+    private EducationData educationData;
 }
