+++ src/main/java/kr/sparta/rchive/domain/educationData/entity/EducationDataTag.java
@@ -1,4 +1,4 @@
-package kr.sparta.rchive.domain.educationdata.entity;
+package kr.sparta.rchive.domain.educationData.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -6,7 +6,7 @@ import jakarta.persistence.IdClass;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import kr.sparta.rchive.domain.user.entity.Track;
+import kr.sparta.rchive.global.entity.BaseTimeEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -18,8 +18,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@IdClass(EducationDataTrackId.class)
-public class EducationDataTrack {
+@IdClass(EducationDataTagId.class)
+public class EducationDataTag extends BaseTimeEntity {
 
     @Id
     @ManyToOne
@@ -28,7 +28,6 @@ public class EducationDataTrack {
 
     @Id
     @ManyToOne
-    @JoinColumn(name = "track_id")
-    private Track track;
-
+    @JoinColumn(name = "tag_id")
+    private Tag tag;
 }
