+++ src/main/java/kr/sparta/rchive/domain/educationData/enums/DataTypeEnum.java
@@ -1,4 +1,4 @@
-package kr.sparta.rchive.domain.educationdata.enums;
+package kr.sparta.rchive.domain.educationData.enums;
 
 public enum DataTypeEnum {
     Video,
